requirements.txt
# Core ML & Data Science
numpy>=1.21.0,<1.24.0
pandas>=1.5.0,<2.1.0
scikit-learn>=1.2.0,<1.4.0

# Deep Learning - TensorFlow 2.10 (compatible with Python 3.10)
tensorflow==2.10.1
keras==2.10.0

# Database
sqlalchemy>=2.0.0,<2.1.0
pymysql>=1.0.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# Data Collection
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
yfinance>=0.2.30

# Scheduling
apscheduler>=3.10.0

# File Processing
openpyxl>=3.1.0
xlrd>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0

# Note: Auto-Keras removed due to dependency conflicts
# We'll use TensorFlow/Keras directly for model training
# This provides the same functionality with better stability
